- src/data-template.js updated: before exporting the data now tabs property for each data item computed.
- Cardsjsx updated:
the addFilter is now functional and passed down to each card, adds one
filter at a time.
removeFilter is functional and passed down to Filter component,
removes one Item if the it passed as a **String** or a batch if they
passed as an **Array**.
the Card component is also provided with tabs prop as defined in data
not computed in this component.

- Card.jsx updated:
Cards with Featured badge also get border-left style.
each tablet rendered now has a key and addCallback to add filter.
propTypes changed as well, now (tabs and addFilter) are required props.

- Card.module.css updated:
the Card now wraps when tabs and details blocks are intercept.
Featured classes added to style Featured Cards with border-left style.
.Name class has same padding as badges to not make the cards
have different width.

- Tablet.jsx updated:
the remove_icon span now registers onClick in filterTablets type to
remove itself from filters.
also the default Tablet type registers addCallback onClick to add itself
to filters.

- Filter.jsx updated:
now it renders nothing if there is no filter as required in the design.
passes props.removeFilter as removeCallback to tablets.
also the #clear button now registers onClick and when clicked, it sends
all filters to remove.

- Filter.module.css updated:
.Tablets class makes the tablets wrap when requred and to display flex
to order them horizontally (fd:row).